requirements.txt
beautifulsoup4==4.13.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3